+++ src/addons/display_commander/ui/new_ui/debug/vulkan_tab.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "../../../ui/imgui_wrapper_base.hpp"
+
+namespace ui::new_ui::debug {
+
+void DrawVulkanTab(display_commander::ui::IImGuiWrapper& imgui);
+
+}  // namespace ui::new_ui::debug
+